Changes made in ant call :
deploy_db  : This ant call is for fresh database creation.
deploy_app : This ant call is for deployment of caTissue Suite application.
deploy_all : This ant call is for deployment and fresh database creation.

upgrade_db  : This ant call is for upgrade the database from caTissue Core 1.2 to caTissue Suite.
upgrade_app : This ant call is for upgrade the deployment of application from caTissue Core 1.2 to caTissue Suite.
upgrade_all : This ant call is for upgrade the deployment of application and also upgrade the database from caTissue Core 1.2 to caTissue Suite.


SVN-Revision: 10959