OPSMN-1981:
It is very common for end users to download bulk import template,
fill columns of interest and upload. End users do not bother to
remove columns, which do not require update, from template.
This means columns that are present but whose values have not
been specified should be treated as absent or should not be
touched.

In view of above, following changes have been implemented:

1. Modified bulk importer to treat empty columns as no change
   columns

2. Introduced ##set_to_blank## to indicate column value should
   be blanked out. This is needed to distinguish between
   columns that should not be touched/modified vs columns whose
   value should be set as empty/blank