OPSMN-2056: Clear UI user state variables on logout
Cause: The UI app keeps track of last state visited by user. This
is helpful in following scenario - an unsigned user visits some
internal URL of app (e.g. clicking link in an e-mail), user is
redirected to sign-in page, on successful sign-in, the user is
redirected back to URL they are interested in. This is all done
using state tracking.

The problem was previously visited state was not cleaned on
logout.

Fix: Clean UI user state variables on logout