OPSMN-2195: Plugin views access mechanism
Implemented plugin views registrar service in core UI app.
The job of registrar is to maintain a mapping of plugin and
their views that should be exposed through various core UI
views.

The plugins during their bootstrap process register custom
views and their access points with registrar

When a core UI view is rendered, the plugin registrar is
consulted to find out whether any custom plugin access view
needs to be rendered. If yes, then its corresponding template
is picked and added to core UI view DOM

It is still responsibility of plugins to define URLs for custom
views and their parent states