OPSMN-2197:
Cause:

Set is used to store event requirements. By definition, Set is
an unordered collection. Therefore when an event is copied,its
associated specimen requirements are copied and created in database
in some random order. The get requirements API uses order of
creation (which is random in case of copy) to sort requirements.

This problem does not exists when explicit sort order is specified
for specimen requirements.

Fix:

When copying event requirements, explicitly set sort order based on
position of requirement during iteration of sorted requirements