OPSMN-2245:
Cause: Views in MySQL do not work as well as they do in Oracle.
Therefore they need a little help in form of cover index.

Fix: Added a covering index consisting of (form context id,
object id and record id) on record entry table to fasten the
process of joining specimen table to appropriate event table
without having to go through each of 1 million specimen event
records.

Although there is index on form context id in form of FK, this
is not used when specimen extension view is used, as it has to
reach out to table to obtain object id and record id details.

(This problem does not exists in Oracle, therefore these all
worked fantabulously well in WashU env that had in excess of 4
million specimen events)