OPSMN-2350: Hand-off authentication context to un-managed thread
Cause: Export of data is done in an un-managed thread. This
un-managed thread has no access to authentication context, which
is used to determine whether row PHI fields require to be masked
or not. This causes NPE.

Fix: At the time of giving task of export to un-managed thread,
hand over authentication context.