OPSMN-2150:
Bootstrap pager requires count of all items in result set in order to
make [prev] and [next] pager button work properly. However doing
counting is not a good idea, as it results in scanning of entire table.
Therefore bulk import pagination uses super-duper idea of fetching one
more item than the no. items required to display (page size) in the view.
That way we trick bootstrap to display/enable [Next] button when no. of
items in result set is more than page size. Since one extra item is fetched,
in controller, we remove the last item (page size + 1) of the result
set. This works fine until we come across a page when no. of items
is less than page size + 1, which is the last page. As a result, the first
BO job is never shown.

Fix: When no. of items obtained is less than page size + 1, do not
remove last item of result set