OPSMN-2563:Oracle: Order data migration fails to migrate order names greater than 255 char There are 2 issues: 1. Datatype change for order name: For this updated the SQL to truncate the order name to 255 char only
2. Migration using order distribution date if order name is null: Oracle
applies the datatype of first record to all the following records. In this
case if the first record's order name is null then it set it as date, for
which it complains later that expected Date but getting char.
For this used concat function to append empty string with date.