OPSMN-2599:
Earlier we had used form.getAllControls() which was giving list of all controls of form including all controls of sub-form and excluded sub-form controls of form. Our schema generation logic is depend on normal controls with sub-form controls. It was generating incorrect csv template because of sub-form control wasn't in the control list.

Fix: Fixed by using form.getOrderedControlList() which is giving only all controls of form including sub-form control but not the controls of sub-form.