OPSMN-2776 Update class & type at parent level using BO. Parent should be stored in a container having restriction other than the type changed. Should give validation error.
Cause: While updating specimen vai BO if update the specimen class or type without specifying specimen location in BO sheet.
As we do partial update of specimen, it changes only specified fields. So if we update specimen class and/or specimen type, existing logic only updates specimen class and/or type. It does not validates container restrictions of specimen position as position is not updating.

Fix: Validate container restrictions if specimen class or type are modified