OPSMN-2970: Read all available entity extensions.
The resource path resolver was given a pattern to stop at first match.
Updated the pattern to get all extension forms. However the last
extension form for a given entity (participant, visit, cp, site or specimen)
overrides the previously found extension forms.

(For now this solves the problem at hand. However, there is a larger
problem to solve, which can't be solved in an hour or so)