OPSMN-3104: Select all and actions performed are now based on selection of specimens displayed in view.
Select-all and actions were based on selection of list specimens.
This was working fine when there was no search/filter specimens
functionality, as the view and model list were same. With implementation of
search, we need to take into account specimens that are displayed in
view rather than the list/cart specimens.