OPSMN-3033: Code review and fix.
The specimens list api was limiting specimens to first 100. This can
result in undesirable behaviour when a shipment/distribution with more
than 100 specimens is executed, as shipment/distribution module/s
depend on specimen list api to obtain specimens which user can
ship/distribute ...

Added criteria parameter to allow list api consumers to specify whether
list specimens api should limit the no. of specimens or not. By default,
the criteria parameter value is false, which is  what is expected at
majority of places.