OPSMN-3145: Added an option in specimen cart UI to allow users add all children/descendants of specimens present in selected cart.
The specimens in cart are now ordered and grouped according to their
ancestors with specimens at root or nearer to root appearing before than
those near the leaf nodes...

Also, added a filter to show only available specimens. This filter is
very useful when users want to select all available specimens and create
a distribution order or shipment.