OPSMN-3195: Participant list view crashes for collection protocols having more than 100 registrations
Cause: While getting visit and specimen count, the query was getting count for all participant's present in the database
due to changes made for list counter, and CPR list was having only 101 items in the list, so while setting count stats,
it was throwing NPE for 102nd registration as its not available.

Solution: While getting visit and specimen count, set the startAt, maxResults and order by id to get count for only those registrations
which are fetched to display in list.