Fixes in wizard widget's forward/backward navigation API.
The widget was using direct function reference to invoke view
supplied callback on forward/backward navigation. This meant
the arguments and their order that could be passed to callback
was fixed. Any view scope arguments could not be passed to
the callback. Moreover this was non-angularish approach.

Fixed by using reference to callback function reference and
materialising the reference in scope context. This allows users
to write callback functions that take both view as well as wizard
supplied parameters as in below example:

scope.validate = function validate(forward, wizard, specimen) {
 //
 // do whatever is required
 //
}

where specimen is view specific object.