OPSMN-3200: To improve performance, by avoiding similar/repetitive container search calls, the search results were cached. For example: searching containers for 5 Buffy Coat aliquots would result in only one call instead of one for each aliquot...
The intention was to cache results for the duration or scope of form;
but inadvertently the search results were cached for lifetime of UI app.
As a result, any newly added containers were not appearing in default
search list.

Fixed the issue by changing scope of caching to lifetime of
form in which the position selector widget appears.