OPSMN-3464: The token stored in browser storage because of previous sign-in makes app think the user is already signed-in. When the user is signed-in, the default app behaviour app is navigate to requested signed-in child view or home view. Later when app learns the token is an expired/stale, it redirects users to sign-in view.
The above behaviour was also triggered for reset password view.
As a result users were seeing sign-in view instead of reset
password view.

Added metadata to states (login and reset password) for which
above default app behaviour should be disabled. This ensures
users see reset password view even when they are signed-in.