OPSMN-3593: Fixed the condition used to display/hide other registered protocols section.
This issue was introduced by fix for OPSMN-3566. The fix was to ensure
'otherCps' attribute in API response doesn't include CP of requested
registration. However, corresponding change in UI was not effected.

Fixed the condition to display other registered protocols when
otherCps is not empty (instead of at least 2).

Added missing translation for i18n key 'participant.other_cps'