OPSMN-3659: When site based access restriction is turned off, all sites are shown in MRN site dropdown. However when the setting is turned on, only those sites for which users have create participant rights are shown in MRN site dropdown.
OPSMN-3661: Ensured "Add to Another Protocol" button is not displayed
when CP list is empty or CP list contains all the CPs for which
participant is already registered.

OPSMN-3662: When site based access restriction is turned off, all
CPs that the user can register participants are considered before
deciding whether to display "Add to Another Protocol" button or not.
When the setting is turned on, CPs based on participant MRN sites
are considered.